+++ UE_AP_state.h
@@ -44,7 +44,7 @@ extern std::vector<UE_node*> UE_node_list;
 extern std::vector<std::vector<double>> Channel_gain_matrix;
 extern std::vector<std::vector<int>> pairing_matrix;
 extern std::vector<std::vector<double>> data_rate_matrix;
-extern std::vector<double> power_allocation_matrix;
+extern std::vector<long double> power_allocation_matrix;
 extern std::vector<int> link_selection_matrix;
 extern std::vector<double> minimum_satisfaction_matrix;
 extern std::vector<double> maximum_requirement_matrix;
